Add validation and replace double primitive type as parameter to string

+++ src/main/java/com/example/demo/domain/entities/PoliticiansRating.java
@@ -20,7 +20,7 @@ public class PoliticiansRating {
 	}
 
 	public double score() {
-		return rating.rating();
+		return rating.rating().doubleValue();
 	}
 
 	public UserRater whoRated() {
